Removes the Status command for license mgmt
This patch substitues the previous Status command with the
the Show command in order to display the license status. This
changes the syntax of the Show command to allow a pattern
show <TYPE> where type is a CollectionByType, and the command
calls the collection getter without any parameters. With this
patch, the command for checking the license status becomes:

show license_status

This patch also fixes a bug in the Install command, where
the command would always return a License object, regardless
of the collection type used in the command.

Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>